Prac 1 finished

+++ WorkPackage1/C/src/CHeterodyning.c
@@ -1,17 +1,18 @@
 #include "CHeterodyning.h"
 
-extern float data [SAMPLE_COUNT];
-extern float carrier[SAMPLE_COUNT];
+
+extern __fp16 data [SAMPLE_COUNT];
+extern __fp16 carrier[SAMPLE_COUNT];
 
 float result [SAMPLE_COUNT];
 
 int main(int argc, char**argv){
     printf("Running Unthreaded Test\n");
     printf("Precision sizeof %ld\n", sizeof(float));
-    
+
 
     printf("Total amount of samples: %ld\n", sizeof(data) / sizeof(data[0]));
-    
+
 
     tic(); // start the timer
     for (int i = 0;i<SAMPLE_COUNT;i++ ){
